BeerService acabado

+++ src/main/java/com/kata_api/beer/repositories/BreweryRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BreweryRepository extends JpaRepository<Brewery, Long> {
+public interface BreweryRepository extends JpaRepository<Brewery, Integer> {
      Brewery findByBreweryId(long breweryId);
      Page<Brewery> findBreweriesByNameContainingIgnoreCase(String breweryName);
      Page<Brewery>findBreweriesByCityContainingIgnoreCase(String city, Pageable pageable);
